Complete refactoring the app class and main.
